requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
camelot-py==0.10.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
cryptography==36.0.1
decorator==5.1.1
distro==1.7.0
dnspython==2.2.1
Elixir==0.7.1
et-xmlfile==1.1.0
Flask==2.0.3
Flask-PyMongo==2.3.0
frozenlist==1.3.0
future==0.18.2
ghostscript==0.7
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
line-bot-sdk==2.1.0
MarkupSafe==2.1.0
multidict==6.0.2
numpy==1.22.2
opencv-python-headless==4.5.5.62
openpyxl==3.0.9
pandas==1.4.1
pbr==5.8.1
pdfminer.six==20211012
pdftopng==0.2.3
pycparser==2.21
pymongo==4.0.2
PyPDF2==1.26.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.2
tabula-py==2.3.0
tabulate==0.8.9
Tempita==0.5.2
urllib3==1.26.8
Werkzeug==2.0.3
xlrd==0.7.1
xlwt==0.7.2
yarl==1.7.2
